Refactor split logic into passes module
